+++ CHANGELOG.md
@@ -10,6 +10,7 @@ These are changes that will probably be included in the next release.
 
 ### Added
  * Include `psutils` to allow some process troubleshooting inside the container
+ * Include custom timescaledb scripts for pgextwlist
 ### Changed
 ### Removed
 ### Fixed